+++ Editor/Utils/CacheLocationValidationResult.cs
@@ -1,8 +1,17 @@
 namespace Unity.AssetManager.Editor
 {
+    internal enum CacheValidationResultError
+    {
+        None,
+        PathTooLong,
+        InvalidPath,
+        CannotWriteToDirectory,
+        DirectoryNotFound
+    }
+    
     internal class CacheLocationValidationResult
     {
-        public bool Success { get; set; }
-        public string Message { get; set; }
+        public bool success { get; set; }
+        public CacheValidationResultError errorType { get; set; }
     }
 }